Minor TextEdit fix, QmlCodeConverter functionality moved to QmlEditFragment
