+++ lib/lveditqmljs/src/qmleditfragment.cpp
@@ -19,10 +19,14 @@
 #include "live/codepalette.h"
 #include "live/projectdocument.h"
 #include "live/projectfile.h"
-#include "qmlcodeconverter.h"
 #include "qmlbindingchannel.h"
 #include "qmlbindingspan.h"
 #include "qmlbindingspanmodel.h"
+#include "live/codeqmlhandler.h"
+#include "live/viewcontext.h"
+
+#include <QJSValue>
+#include <QJSValueIterator>
 
 namespace lv{
 
@@ -81,6 +85,26 @@ QmlEditFragment::~QmlEditFragment(){
 
     delete m_bindingSpanModel;
     delete m_bindingSpan;
+}
+
+QObject *QmlEditFragment::createObject(const DocumentQmlInfo::ConstPtr &info, const QString &declaration, const QString &path, QObject *parent)
+{
+    QString fullDeclaration;
+
+    for ( auto it = info->imports().begin(); it != info->imports().end(); ++it ){
+        fullDeclaration +=
+            "import " + it->uri() + " " +
+            QString::number(it->versionMajor()) + "." + QString::number(it->versionMinor());
+        if ( !it->as().isEmpty() ){
+            fullDeclaration += " as " + it->as();
+        }
+        fullDeclaration += "\n";
+    }
+    fullDeclaration += "\n" + declaration + "\n";
+
+    QObject* obj = ViewContext::instance().engine()->createObject(fullDeclaration, parent, path);
+
+    return obj;
 }
 
 void QmlEditFragment::setRelativeBinding(const QSharedPointer<QmlBindingPath> &bp){
@@ -235,10 +259,97 @@ QString QmlEditFragment::objectId()
     return m_objectId;
 }
 
+void QmlEditFragment::writeProperties(const QJSValue &properties)
+{
+    if ( !properties.isObject() ){
+        qWarning("Properties must be of object type, use 'write' to add code directly.");
+        return;
+    }
+
+    int valuePosition = declaration()->valuePosition();
+    int valueLength   = declaration()->valueLength();
+    QString source    = declaration()->document()->content().mid(valuePosition, valueLength);
+
+    lv::DocumentQmlInfo::Ptr docinfo = lv::DocumentQmlInfo::create(declaration()->document()->file()->path());
+    if ( !docinfo->parse(source) )
+        return;
+
+    lv::DocumentQmlValueObjects::Ptr objects = docinfo->createObjects();
+    lv::DocumentQmlValueObjects::RangeObject* root = objects->root();
+
+    QSet<QString> leftOverProperties;
+
+    QJSValueIterator it(properties);
+    while ( it.hasNext() ){
+        it.next();
+        leftOverProperties.insert(it.name());
+    }
+
+    QString indent = "";
+
+    for ( auto it = root->properties.begin(); it != root->properties.end(); ++it ){
+        lv::DocumentQmlValueObjects::RangeProperty* p = *it;
+        QString propertyName = source.mid(p->begin, p->propertyEnd - p->begin);
+
+        if ( indent.isEmpty() ){
+            int index = p->begin - 1;
+            while (index >= 0) {
+                if ( source[index] == QChar('\n') ){
+                    break;
+                }
+                if ( !source[index].isSpace() )
+                    break;
+                --index;
+            }
+            if ( index + 1 < p->begin - 1){
+                indent = source.mid(index + 1, p->begin - 1 - index);
+            }
+        }
+
+        if ( leftOverProperties.contains(propertyName) ){
+            source.replace(p->valueBegin, p->end - p->valueBegin, buildCode(properties.property(propertyName)));
+            leftOverProperties.remove(propertyName);
+        }
+    }
+
+    int writeIndex = source.length() - 2;
+    while ( writeIndex >= 0 ){
+        if ( !source[writeIndex].isSpace() ){
+            break;
+        }
+        --writeIndex;
+    }
+
+    if ( indent.isEmpty() ){
+        int indentIndex = source.length() - 2;
+        while ( indentIndex >= 0 ){
+            if ( source[indentIndex] == QChar('\n') ){
+                break;
+            }
+            if ( !source[indentIndex].isSpace() )
+                break;
+            --indentIndex;
+        }
+        if ( indentIndex + 1 < source.length() - 2){
+            indent = source.mid(indentIndex + 1, source.length() - 2 - indentIndex) + "    ";
+        }
+    }
+
+    for ( auto it = leftOverProperties.begin(); it != leftOverProperties.end(); ++it ){
+        source.insert(writeIndex + 1, "\n" + indent + *it + ": " + buildCode(properties.property(*it)));
+    }
+
+    write(source);
+}
+
+void QmlEditFragment::write(const QJSValue value){
+    writeCode(buildCode(value));
+}
+
 /**
  * \brief Writes the \p code to the value part of this fragment
  */
-void QmlEditFragment::write(const QString &code){
+void QmlEditFragment::writeCode(const QString &code){
     ProjectDocument* document = m_declaration->document();
 
     if ( document->editingStateIs(ProjectDocument::Palette))
@@ -295,6 +406,61 @@ QObject *QmlEditFragment::readObject()
 
     return nullptr;
 }
+
+QVariant QmlEditFragment::parse()
+{
+    QString val = readValueText();
+    if ( val.length() > 1 && (
+             (val.startsWith('"') && val.endsWith('"')) ||
+             (val.startsWith('\'') && val.endsWith('\'') )
+        ))
+    {
+        return QVariant(val.mid(1, val.length() - 2));
+    } else if ( val == "true" ){
+        return QVariant(true);
+    } else if ( val == "false" ){
+        return QVariant(false);
+    } else {
+        bool ok;
+        qint64 number = val.toLongLong(&ok);
+        if (ok)
+            return QVariant(number);
+
+        return QVariant(val.toDouble());
+    }
+}
+
+void QmlEditFragment::updateBindings()
+{
+    if (bindingPalette())
+        m_whenBinding.call();
+}
+
+void QmlEditFragment::updateFromPalette()
+{
+    CodePalette* palette = qobject_cast<CodePalette*>(sender());
+    if ( palette && bindingSpan() ){
+        bindingSpan()->commit(palette->value());
+    }
+}
+
+void QmlEditFragment::suggestionsForExpression(const QString &expression, CodeCompletionModel *model, bool suggestFunctions)
+{
+    QObject* editParent = parent();
+    CodeQmlHandler* qmlHandler = nullptr;
+    while ( editParent ){
+        qmlHandler = qobject_cast<CodeQmlHandler*>(editParent);
+        if ( qmlHandler )
+            break;
+
+        editParent = editParent->parent();
+    }
+
+    if (qmlHandler){
+        qmlHandler->suggestionsForProposedExpression(declaration(), expression, model, suggestFunctions);
+    }
+}
+
 
 /**
  * \brief Reads the code value of this fragment and returns it.
@@ -415,8 +581,7 @@ void QmlEditFragment::updateValue(){
         }
         if ( m_bindingPalette ){
             m_bindingPalette->setValueFromBinding(inputPath->property().read());
-            QmlCodeConverter* cvt = static_cast<QmlCodeConverter*>(m_bindingPalette->extension());
-            cvt->whenBinding().call();
+            m_whenBinding.call();
         }
     }
 }
@@ -480,5 +645,112 @@ void QmlEditFragment::signalObjectAdded(QmlEditFragment *ef, QPointF cursorCoord
     emit objectAdded(ef, cursorCoords);
 }
 
+
+bool QmlEditFragment::bindExpression(const QString &expression)
+{
+    QObject* editParent = parent();
+    CodeQmlHandler* qmlHandler = nullptr;
+    while ( editParent ){
+        qmlHandler = qobject_cast<CodeQmlHandler*>(editParent);
+        if ( qmlHandler )
+            break;
+
+        editParent = editParent->parent();
+    }
+
+    if (qmlHandler){
+        return qmlHandler->findBindingForExpression(this, expression);
+    }
+
+    return false;
+}
+
+bool QmlEditFragment::bindFunctionExpression(const QString &expression){
+    QObject* editParent = parent();
+    CodeQmlHandler* qmlHandler = nullptr;
+    while ( editParent ){
+        qmlHandler = qobject_cast<CodeQmlHandler*>(editParent);
+        if ( qmlHandler )
+            break;
+
+        editParent = editParent->parent();
+    }
+
+    if (qmlHandler){
+        return qmlHandler->findFunctionBindingForExpression(this, expression);
+    }
+
+    return false;
+}
+
+QString QmlEditFragment::buildCode(const QJSValue &value){
+    if ( value.isArray() ){
+        QString result;
+        QJSValueIterator it(value);
+        result = "[";
+        bool first = true;
+        while ( it.hasNext() ){
+            it.next();
+            if ( it.name() != "length" ){
+                if ( first ){
+                    first = false;
+                } else {
+                    result += ",";
+                }
+                result += buildCode(it.value());
+            }
+        }
+        result += "]";
+        return result;
+    } else if ( value.isDate() ){
+        QDateTime dt = value.toDateTime();
+
+        QString year    = QString::number(dt.date().year()) + ",";
+        QString month   = QString::number(dt.date().month()) + ",";
+        QString day     = QString::number(dt.date().day()) + ",";
+        QString hour    = QString::number(dt.time().hour()) + ",";
+        QString minute  = QString::number(dt.time().minute()) + ",";
+        QString second  = QString::number(dt.time().second()) + ",";
+        QString msecond = QString::number(dt.time().msec());
+
+        return "new Date(" + year + month + day + hour + minute + second + msecond + ")";
+
+    } else if ( value.isObject() ){
+        if ( value.hasOwnProperty("__ref") ){
+            QString result = value.property("__ref").toString();
+            if ( result.isEmpty() ){
+                return "null";
+            } else {
+                return result;
+            }
+        }
+
+        QString text = value.toString();
+        if (text.length() > 6 && text.left(6) == "QSizeF"){
+            return "'" + value.property("width").toString() + "x" + value.property("height").toString() + "'";
+        }
+
+        QString result = "{";
+        bool first = true;
+        QJSValueIterator it(value);
+        while ( it.hasNext() ){
+            it.next();
+            if ( first ){
+                first = false;
+            } else {
+                result += ",";
+            }
+
+            result += it.name() + ": " + buildCode(it.value());
+        }
+
+        return result + "}";
+    } else if ( value.isString() ){
+        return "'" + value.toString() + "'";
+    } else if ( value.isBool() || value.isNumber() ){
+        return value.toString();
+    }
+    return "";
+}
 
 }// namespace
