+++ plugins/editor/palettes/BooleanPalette.qml
@@ -38,7 +38,7 @@ CodePalette {
     onCheckedChanged: {
         palette.value = checked
         if ( !palette.isBindingChange() ){
-            extension.write(palette.value)
+            editFragment.write(palette.value)
         }
     }
 
@@ -46,9 +46,9 @@ CodePalette {
         checked = value
     }
 
-    onExtensionChanged: {
-        extension.whenBinding = function(){
-            extension.write(palette.value)
+    onEditFragmentChanged: {
+        editFragment.whenBinding = function(){
+            editFragment.write(palette.value)
         }
     }
