+++ plugins/editor/palettes/DoubleHistoryPlotPalette.qml
@@ -28,9 +28,9 @@ CodePalette{
         }
     }
 
-    onExtensionChanged: {
-        extension.whenBinding = function(){
-            extension.write(palette.value)
+    onEditFragmentChanged: {
+        editFragment.whenBinding = function(){
+            editFragment.write(palette.value)
         }
     }
