+++ plugins/editor/palettes/DoublePalette.qml
@@ -59,7 +59,7 @@ CodePalette{
                 roundValue = roundValue.toFixed(2)
                 palette.value = roundValue
                 if ( !palette.isBindingChange() )
-                    extension.write(palette.value)
+                    editFragment.write(palette.value)
             }
             stepSize: 1.0
             maximumValue: 25
@@ -100,7 +100,7 @@ CodePalette{
                     palette.value = roundValue
 
                 if ( !palette.isBindingChange() )
-                    extension.write(palette.value)
+                    editFragment.write(palette.value)
             }
             stepSize: 0.01
             maximumValue: 0.99
@@ -221,9 +221,9 @@ CodePalette{
 
     }
 
-    onExtensionChanged: {
-        extension.whenBinding = function(){
-            extension.write(palette.value)
+    onEditFragmentChanged: {
+        editFragment.whenBinding = function(){
+            editFragment.write(palette.value)
         }
     }
