+++ plugins/editor/palettes/ImportsPalette.qml
@@ -177,9 +177,9 @@ CodePalette{
         }
     }
 
-    onExtensionChanged: {
-        extension.whenBinding = function(){
-            extension.write(palette.value)
+    onEditFragmentChanged: {
+        editFragment.whenBinding = function(){
+            editFragment.write(palette.value)
         }
     }
 }
