+++ plugins/editor/palettes/IntHistoryPlotPalette.qml
@@ -21,9 +21,9 @@ CodePalette{
         valueHistory.currentValue = value
     }
 
-    onExtensionChanged: {
-        extension.whenBinding = function(){
-            extension.write(palette.value)
+    onEditFragmentChanged: {
+        editFragment.whenBinding = function(){
+            editFragment.write(palette.value)
         }
     }
 }
