+++ plugins/lcvcore/palettes/VideoCapturePalette.qml
@@ -33,9 +33,9 @@ CodePalette{
 
         Connections{
             target: captureContainer.videoCapture
-            onFpsChanged: extension.updateBindings()
-            onCurrentFrameChanged: extension.updateBindings()
-            onPausedChanged: extension.updateBindings()
+            onFpsChanged: editFragment.updateBindings()
+            onCurrentFrameChanged: editFragment.updateBindings()
+            onPausedChanged: editFragment.updateBindings()
         }
 
         width: 290
@@ -53,7 +53,7 @@ CodePalette{
                 if ( !parent || !parent.videoCapture )
                     return
                 parent.videoCapture.fps = fpsSlider.value
-                extension.writeProperties({
+                editFragment.writeProperties({
                     'fps' : fpsSlider.value
                 })
             }
@@ -82,12 +82,12 @@ CodePalette{
             videoCapture : parent.videoCapture
             width: 290
             onSeekTriggered: {
-                extension.writeProperties({
+                editFragment.writeProperties({
                     'currentFrame' : videoCapture.currentFrame
                 })
             }
             onPlayPauseTriggered: {
-                extension.writeProperties({
+                editFragment.writeProperties({
                     'paused' : videoCapture.paused
                 })
             }
@@ -100,9 +100,9 @@ CodePalette{
         captureContainer.videoCapture = value
     }
 
-    onExtensionChanged: {
-        extension.whenBinding = function(){
-            extension.writeProperties({
+    onEditFragmentChanged: {
+        editFragment.whenBinding = function(){
+            editFragment.writeProperties({
                 'fps' : palette.value.fps,
                 'currentFrame' : palette.value.currentFrame,
                 'paused' : palette.value.paused
